# artifacts/api-server/src/routes/grocery.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { groceryListsTable, groceryItemsTable, propertiesTable } from "@workspace/db";
import { eq, sql, inArray } from "drizzle-orm";
import { getApprovedHouseholdScope } from "../middlewares/requireApprovedHousehold";

const router = Router();

router.get("/grocery-lists", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    if (scope.propertyIds.length === 0) {
      res.json([]);
      return;
    }

    const lists = await db
      .select({
        list: groceryListsTable,
        propertyName: propertiesTable.name,
        itemCount: sql<number>`count(${groceryItemsTable.id})::int`,
        checkedCount: sql<number>`count(case when ${groceryItemsTable.checked} = true then 1 end)::int`,
      })
      .from(groceryListsTable)
      .leftJoin(propertiesTable, eq(groceryListsTable.propertyId, propertiesTable.id))
      .leftJoin(groceryItemsTable, eq(groceryItemsTable.listId, groceryListsTable.id))
      .where(inArray(groceryListsTable.propertyId, scope.propertyIds))
      .groupBy(groceryListsTable.id, propertiesTable.name)
      .orderBy(groceryListsTable.id);

    res.json(
      lists.map((r) => ({
        id: String(r.list.id),
        name: r.list.name,
        propertyId: String(r.list.propertyId),
        propertyName: r.propertyName ?? "",
        itemCount: r.itemCount,
        checkedCount: r.checkedCount,
        createdAt: r.list.createdAt.toISOString(),
      })),
    );
  } catch (err) {
    req.log.error({ err }, "Failed to get grocery lists");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/grocery-lists", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    const { name, propertyId } = req.body;
    if (!name || !propertyId) {
      res.status(400).json({ error: "name and propertyId required" });
      return;
    }

    const propertyIdNum = Number(propertyId);
    if (!Number.isInteger(propertyIdNum)) {
      res.status(400).json({ error: "propertyId must be a valid id" });
      return;
    }
    if (!scope.propertyIds.includes(propertyIdNum)) {
      res.status(403).json({ error: "Property not accessible" });
      return;
    }

    const [list] = await db
      .insert(groceryListsTable)
      .values({ name, propertyId: propertyIdNum })
      .returning();

    const [prop] = await db.select().from(propertiesTable).where(eq(propertiesTable.id, list.propertyId));

    res.status(201).json({
      id: String(list.id),
      name: list.name,
      propertyId: String(list.propertyId),
      propertyName: prop?.name ?? "",
      itemCount: 0,
      checkedCount: 0,
      createdAt: list.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Failed to create grocery list");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.delete("/grocery-lists/:id", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    const listId = Number(req.params.id);
    if (!Number.isInteger(listId)) {
      res.status(400).json({ error: "Invalid list id" });
      return;
    }

    const [list] = await db
      .select()
      .from(groceryListsTable)
      .where(eq(groceryListsTable.id, listId))
      .limit(1);
    if (!list || !scope.propertyIds.includes(list.propertyId)) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    await db.delete(groceryListsTable).where(eq(groceryListsTable.id, listId));
    res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Failed to delete grocery list");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/grocery-lists/:id/items", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    const listId = Number(req.params.id);
    if (!Number.isInteger(listId)) {
      res.status(400).json({ error: "Invalid list id" });
      return;
    }

    const [list] = await db
      .select()
      .from(groceryListsTable)
      .where(eq(groceryListsTable.id, listId))
      .limit(1);
    if (!list || !scope.propertyIds.includes(list.propertyId)) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    const items = await db
      .select()
      .from(groceryItemsTable)
      .where(eq(groceryItemsTable.listId, listId))
      .orderBy(groceryItemsTable.createdAt);

    res.json(
      items.map((i) => ({
        id: String(i.id),
        listId: String(i.listId),
        name: i.name,
        quantity: i.quantity ?? null,
        category: i.category ?? null,
        checked: i.checked,
        addedBy: i.addedBy ?? null,
        createdAt: i.createdAt.toISOString(),
      })),
    );
  } catch (err) {
    req.log.error({ err }, "Failed to get grocery items");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/grocery-lists/:id/items", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    const listId = Number(req.params.id);
    if (!Number.isInteger(listId)) {
      res.status(400).json({ error: "Invalid list id" });
      return;
    }
    const { name, quantity, category, addedBy } = req.body;
    if (!name) {
      res.status(400).json({ error: "name required" });
      return;
    }

    const [list] = await db
      .select()
      .from(groceryListsTable)
      .where(eq(groceryListsTable.id, listId))
      .limit(1);
    if (!list || !scope.propertyIds.includes(list.propertyId)) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    const [item] = await db
      .insert(groceryItemsTable)
      .values({ listId, name, quantity: quantity ?? null, category: category ?? null, addedBy: addedBy ?? null })
      .returning();

    res.status(201).json({
      id: String(item.id),
      listId: String(item.listId),
      name: item.name,
      quantity: item.quantity ?? null,
      category: item.category ?? null,
      checked: item.checked,
      addedBy: item.addedBy ?? null,
      createdAt: item.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Failed to add grocery item");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.put("/grocery-items/:id", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    const id = Number(req.params.id);
    if (!Number.isInteger(id)) {
      res.status(400).json({ error: "Invalid item id" });
      return;
    }
    const { name, quantity, category, checked } = req.body;

    const [existing] = await db
      .select({ item: groceryItemsTable, propertyId: groceryListsTable.propertyId })
      .from(groceryItemsTable)
      .innerJoin(groceryListsTable, eq(groceryItemsTable.listId, groceryListsTable.id))
      .where(eq(groceryItemsTable.id, id))
      .limit(1);
    if (!existing || !scope.propertyIds.includes(existing.propertyId)) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    await db
      .update(groceryItemsTable)
      .set({
        ...(name !== undefined && { name }),
        ...(quantity !== undefined && { quantity: quantity ?? null }),
        ...(category !== undefined && { category: category ?? null }),
        ...(checked !== undefined && { checked }),
      })
      .where(eq(groceryItemsTable.id, id));

    const [item] = await db.select().from(groceryItemsTable).where(eq(groceryItemsTable.id, id));
    if (!item) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    res.json({
      id: String(item.id),
      listId: String(item.listId),
      name: item.name,
      quantity: item.quantity ?? null,
      category: item.category ?? null,
      checked: item.checked,
      addedBy: item.addedBy ?? null,
      createdAt: item.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Failed to update grocery item");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.delete("/grocery-items/:id", async (req, res) => {
  try {
    const scope = getApprovedHouseholdScope(res);
    const id = Number(req.params.id);
    if (!Number.isInteger(id)) {
      res.status(400).json({ error: "Invalid item id" });
      return;
    }

    const [existing] = await db
      .select({ item: groceryItemsTable, propertyId: groceryListsTable.propertyId })
      .from(groceryItemsTable)
      .innerJoin(groceryListsTable, eq(groceryItemsTable.listId, groceryListsTable.id))
      .where(eq(groceryItemsTable.id, id))
      .limit(1);
    if (!existing || !scope.propertyIds.includes(existing.propertyId)) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    await db.delete(groceryItemsTable).where(eq(groceryItemsTable.id, id));
    res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Failed to delete grocery item");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;