# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HomeHub household management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grocery-lists:
    get:
      operationId: getGroceryLists
      tags: [grocery]
      summary: List all grocery lists
      responses:
        "200":
          description: List of grocery lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryList"
    post:
      operationId: createGroceryList
      tags: [grocery]
      summary: Create a grocery list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroceryListInput"
      responses:
        "201":
          description: Created grocery list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
  /grocery-lists/{id}:
    delete:
      operationId: deleteGroceryList
      tags: [grocery]
      summary: Delete a grocery list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /grocery-lists/{id}/items:
    get:
      operationId: getGroceryItems
      tags: [grocery]
      summary: Get items in a grocery list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Grocery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryItem"
    post:
      operationId: addGroceryItem
      tags: [grocery]
      summary: Add item to grocery list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroceryItemInput"
      responses:
        "201":
          description: Created grocery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"
  /grocery-items/{id}:
    put:
      operationId: updateGroceryItem
      tags: [grocery]
      summary: Update a grocery item (check/uncheck, rename)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroceryItemInput"
      responses:
        "200":
          description: Updated grocery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"
    delete:
      operationId: deleteGroceryItem
      tags: [grocery]
      summary: Delete a grocery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    GroceryList:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        propertyId:
          type: string
        propertyName:
          type: string
        itemCount:
          type: integer
        checkedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - propertyId
        - propertyName
        - itemCount
        - checkedCount
        - createdAt
    CreateGroceryListInput:
      type: object
      properties:
        name:
          type: string
        propertyId:
          type: string
      required:
        - name
        - propertyId
    GroceryItem:
      type: object
      properties:
        id:
          type: string
        listId:
          type: string
        name:
          type: string
        quantity:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        checked:
          type: boolean
        addedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - listId
        - name
        - checked
        - createdAt
    CreateGroceryItemInput:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        addedBy:
          type: string
          nullable: true
      required:
        - name
    UpdateGroceryItemInput:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        checked:
          type: boolean
...